[FIX] website_blog wysiwyg
