+++ addons/website_blog/views/website_blog_templates.xml
@@ -3,6 +3,8 @@
 
 <template id="assets_wysiwyg" inherit_id="website.assets_wysiwyg" name="Blog posts snippet options">
     <xpath expr="." position="inside">
+        <script type="text/javascript" src="/website_blog/static/src/js/options.js"></script>
+        <script type="text/javascript" src="/website_blog/static/src/js/wysiwyg.js"></script>
         <script type="text/javascript" src="/website_blog/static/src/snippets/s_latest_posts/options.js"/>
     </xpath>
 </template>
